# server/swagger.yaml
openapi: 3.0.0
info:
  title: Blog API
  version: 1.0.0
  description: API documentation for the Blog application
servers:
  - url: http://localhost:4000
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
          minLength: 2
          maxLength: 50
        email:
          type: string
          format: email
        role:
          type: string
          enum: [user, author, admin]
          default: user
        bio:
          type: string
          maxLength: 500
        isAccountVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Comment:
      type: object
      properties:
        _id:
          type: string
        content:
          type: string
        post:
          type: string
        author:
          $ref: "#/components/schemas/User"
        parentComment:
          type: string
          nullable: true
        likes:
          type: array
          items:
            type: string
        isEdited:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
...
paths:
...
  /api/posts/{postId}/comments:
    get:
      tags:
        - Comments
      summary: Get comments for a post (with pagination)
      parameters:
        - in: path
          name: postId
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
          description: Number of comments per page
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Comment"
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      pages:
                        type: integer
    post:
      tags:
        - Comments
      summary: Add a comment to a post
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: postId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                parentComment:
                  type: string
                  nullable: true
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Comment"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...